feat: take a u64 total in PaginationResponse::new

+++ crates/tycho-indexer/src/services/rpc.rs
@@ -325,11 +325,7 @@ where
                 .into_iter()
                 .map(dto::ResponseAccount::from)
                 .collect(),
-            PaginationResponse::new(
-                pagination_params.page,
-                pagination_params.page_size,
-                total as i64,
-            ),
+            PaginationResponse::new(pagination_params.page, pagination_params.page_size, total),
         ))
     }
 
@@ -497,14 +493,14 @@ where
         // By precomputing the paginated IDs we also ensure that the fetched balances and
         // protocol state are paginated in the same way.
         // Also, by doing this in a single point prevents failures and increases performance.
-        let (paginated_ids, total): (Vec<String>, i64) = match ids {
+        let (paginated_ids, total): (Vec<String>, u64) = match ids {
             Some(ids) => (
                 ids.iter()
                     .skip(pagination_params.offset() as usize)
                     .take(pagination_params.page_size as usize)
                     .map(|s| s.to_string())
                     .collect(),
-                ids.len() as i64,
+                ids.len() as u64,
             ),
             None => {
                 let req = dto::ProtocolComponentsRequestBody {
@@ -518,7 +514,7 @@ where
                     .get_protocol_components_inner(req)
                     .await
                     .expect("Failed to get protocol component IDs");
-                let total_components = protocol_components.pagination.total;
+                let total_components = protocol_components.pagination.total as u64;
                 (
                     protocol_components
                         .protocol_components
@@ -595,7 +591,7 @@ where
                 .collect(),
             None => self.protocol_systems.iter().collect(),
         };
-        let total = filtered.len() as i64;
+        let total = filtered.len() as u64;
         let page = request.pagination.page;
         let page_size = request.pagination.page_size;
         let skip = (page * page_size) as usize;
@@ -653,7 +649,7 @@ where
                 PaginationResponse::new(
                     pagination_params.page,
                     pagination_params.page_size,
-                    tvl.total.unwrap_or_default() as i64,
+                    tvl.total.unwrap_or_default(),
                 ),
             )),
             Err(err) => {
@@ -728,7 +724,7 @@ where
                 &PaginationResponse::new(
                     request.pagination.page,
                     request.pagination.page_size,
-                    token_data.total.unwrap_or_default() as i64,
+                    token_data.total.unwrap_or_default(),
                 ),
             )),
             Err(err) => {
@@ -806,7 +802,7 @@ where
                 .map(|comp| comp.id.as_str())
                 .collect();
 
-            let total = buffered_components.len() as i64;
+            let total = buffered_components.len() as u64;
 
             if requested_ids.len() == fetched_ids.len() {
                 let response_components: Vec<dto::ProtocolComponent> = buffered_components
@@ -878,7 +874,7 @@ where
                     PaginationResponse::new(
                         pagination_params.page,
                         pagination_params.page_size,
-                        total as i64,
+                        total,
                     ),
                 ))
             }
@@ -1027,7 +1023,7 @@ where
                 request.pagination.page_size,
                 entry_points_tracing_params_data
                     .total
-                    .unwrap_or_default() as i64,
+                    .unwrap_or_default(),
             ),
         })
     }
